Improving colors and lights and starting shadows

+++ srcs/transformations.c
@@ -19,8 +19,8 @@ static t_vec3	ft_rotation_z(t_vec3 ex_pos, float angle)
 	t_vec3	new_pos;
 
 	angle = ft_deg2rad(angle);
-	new_x = ex_pos.x * cos(angle) + ex_pos.y * -sin(angle);
-	new_y = ex_pos.x * sin(angle) + ex_pos.y * cos(angle);
+	new_x = ex_pos.x * cosf(angle) + ex_pos.y * -sinf(angle);
+	new_y = ex_pos.x * sinf(angle) + ex_pos.y * cosf(angle);
 	new_pos.x = new_x;
 	new_pos.y = new_y;
 	new_pos.z = ex_pos.z;
@@ -34,8 +34,8 @@ static t_vec3	ft_rotation_y(t_vec3 ex_pos, float angle)
 	t_vec3	new_pos;
 
 	angle = ft_deg2rad(angle);
-	new_x = ex_pos.x * cos(angle) + ex_pos.z * sin(angle);
-	new_z = ex_pos.x * -sin(angle) + ex_pos.z * cos(angle);
+	new_x = ex_pos.x * cosf(angle) + ex_pos.z * sinf(angle);
+	new_z = ex_pos.x * -sinf(angle) + ex_pos.z * cosf(angle);
 	new_pos.x = new_x;
 	new_pos.z = new_z;
 	new_pos.y = ex_pos.y;
@@ -49,8 +49,8 @@ static t_vec3	ft_rotation_x(t_vec3 ex_pos, float angle)
 	t_vec3	new_pos;
 
 	angle = ft_deg2rad(angle);
-	new_y = ex_pos.y * cos(angle) + ex_pos.z * -sin(angle);
-	new_z = ex_pos.y * sin(angle) + ex_pos.z * cos(angle);
+	new_y = ex_pos.y * cosf(angle) + ex_pos.z * -sinf(angle);
+	new_z = ex_pos.y * sinf(angle) + ex_pos.z * cosf(angle);
 	new_pos.y = new_y;
 	new_pos.z = new_z;
 	new_pos.x = ex_pos.x;
@@ -79,7 +79,6 @@ void			transformations(t_vec3 *position, char *str, int *j)
 	loop(str, j);
 	rotation_degree = (t_vec3){0, 0, 0};
 	rotation_degree = assign_vectors(str, j, rotation_degree);
-	//rotation_degree = atoi_custom(str, j);
 	*position = ft_rotation_x(*position, rotation_degree.x);
 	*position = ft_rotation_y(*position, rotation_degree.y);
 	*position = ft_rotation_z(*position, rotation_degree.z);
